Add UrlRewriteFilter to remove .html from end of file, move Trip.java and TripDay.java into the data folder, and remove the trips files from the separate folder.
