+++ src/main/java/com/google/sps/servlets/AuthServlet.java
@@ -35,7 +35,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/auth")
 public class AuthServlet extends HttpServlet {
 
   // Set redirect URL after login / logout as index page.
