+++ src/main/java/com/google/sps/servlets/UserTripServlet.java
@@ -28,7 +28,7 @@ import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
 import com.google.sps.data.User;
-import com.google.sps.Trip;
+import com.google.sps.data.Trip;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +37,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/user-trips")
 public class UserTripServlet extends HttpServlet {
 
   // This allows the doGet method to return null.
