+++ src/main/webapp/calendar.html
@@ -29,7 +29,7 @@
               <a class="nav-link" href="/">Create New Trip</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/trips/">Trips</a>
+              <a class="nav-link" href="/trips">Trips</a>
             </li>
             <li class="nav-item">
               <a class="nav-link" id="sign-out-link">Sign out</a>
