+++ src/main/webapp/createNavLinks.js
@@ -32,7 +32,7 @@ function createGroupNavLinks() {
   const tripKeyQuery = (tripKey != null && tripKey != '') ? '?tripKey=' + tripKey : '';
 
   // Add the href links to the buttons.
-  calendarButton.href = '/calendar.html' + tripKeyQuery;
-  mapsButton.href = '/maps.html' + tripKeyQuery;
-  editButton.href = '/edit.html' + tripKeyQuery;
+  calendarButton.href = '/calendar' + tripKeyQuery;
+  mapsButton.href = '/maps' + tripKeyQuery;
+  editButton.href = '/edit' + tripKeyQuery;
 }
