+++ src/main/webapp/index.html
@@ -28,7 +28,7 @@
               <a class="nav-link" href="/">Create New Trip <span class="sr-only">(current)</span></a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/trips/">Trips</a>
+              <a class="nav-link" href="/trips">Trips</a>
             </li>
             <li class="nav-item">
               <a class="nav-link" id="sign-out-link">Sign out</a>
