+++ src/test/java/com/google/sps/CalendarServletTest.java
@@ -28,8 +28,9 @@ import com.google.appengine.api.datastore.Query.FilterPredicate;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
 import com.google.sps.data.Event;
+import com.google.sps.data.Trip;
+import com.google.sps.data.TripDay;
 import com.google.sps.data.User;
-import com.google.sps.Trip;
 import com.google.sps.servlets.CalendarServlet;
 import java.io.PrintWriter;
 import java.io.StringWriter;
