+++ src/test/java/com/google/sps/MapServletTest.java
@@ -22,7 +22,8 @@ import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
-import com.google.sps.Trip;
+import com.google.sps.data.Trip;
+import com.google.sps.data.TripDay;
 import com.google.sps.data.User;
 import com.google.sps.servlets.AuthServlet;
 import com.google.sps.servlets.MapServlet;
