+++ src/test/java/com/google/sps/TripServletTest.java
@@ -42,8 +42,9 @@ import com.google.maps.model.FindPlaceFromText;
 import com.google.maps.model.PlaceDetails;
 import com.google.maps.model.PlacesSearchResult;
 import com.google.maps.model.TravelMode;
-import com.google.sps.Trip;
 import com.google.sps.data.Event;
+import com.google.sps.data.Trip;
+import com.google.sps.data.TripDay;
 import com.google.sps.data.User;
 import com.google.sps.servlets.AuthServlet;
 import com.google.sps.servlets.TripServlet;
